Program should work even without global configuration
One of the features should be that fmakegen can work
even though global configuration does not exist or
is damaged. This to make sure fmakegen is somewhat
usable even without installation.

Added printWarm to give warnings about configure not working.

Changed .ini to .cfg as libconfig file format is not INI.

Added missing close.

On branch master
Changes to be committed:
	modified:   Src/Main.c